Update contract.ts
In this modified version, when it is the user's first donation, the storage cost (STORAGE_COST) is subtracted from the toTransfer amount and added to the donatedSoFar variable. This ensures that the fee is not transferred to the beneficiary and is accounted for as part of the user's total donation amount.